Support mutable visitation
The old visitor did not allow mutating the AST as it traversed it, while
this one does.

This visitor is becoming increasingly hard to read/maintain, but I think
this macro-based approach is still better than maintaining the immutable
and mutable copies separately. The real solution is a procedural macro
that automatically generates visitors at compile time, but that's quite
a bit of work.
