+++ src/ast/mod.rs
@@ -17,7 +17,26 @@ mod ddl;
 mod operator;
 mod query;
 mod value;
-pub mod visit;
+#[macro_use]
+mod visit_macro;
+
+pub mod visit {
+    // Visiting an enum struct variant with many arguments forces our hand here.
+    // If, over time, we convert these large variants to dedicated structs, we
+    // can remove this escape hatch.
+    #![allow(clippy::too_many_arguments)]
+    // Disable lints that want us to rewrite `&Ident` as `&str`, as `&str` is not as
+    // self-documenting as &Ident.
+    #![allow(clippy::ptr_arg)]
+    make_visitor!(Visit: &);
+}
+
+pub mod visit_mut {
+    // See justification for these attributes in the `visit` module.
+    #![allow(clippy::too_many_arguments)]
+    #![allow(clippy::ptr_arg)]
+    make_visitor!(VisitMut: &mut);
+}
 
 use std::fmt;
